Cleaning up
Edits to board() and room() to make it look better. Fights are almost
finished, except cards share health so any attack against one wolf card
is felt on every wolf card. Need to figure out a way to fix that. The
last idea was to change how the stagnant player deck was filled, where
instead of being set to allCard at a value, instead it would be more
specific, allCard.get(8).getCardHealth()... I don't know it it works.

If you switch 8 with a changing variable, then when filling arrays it
would make sense to make a preset array full of just the directions from
the allacard Array, but I don't know